MultipleSelectItems:
Now, when selecting multiple game objects in LevelEditor, if they have all the same component types, all can be modified. For example, all the game objetcs that are floor have to have rigid bodies. Now, we can select all the game objects that are pretending to be a floor and put RigidBody2d component with just one click.